fix(hitbar) : 2023. 10 Survivor ActionManager에서 hitbar 부분 분리 중
uiManager라는 스크립트를 만들고 추후에 lateUpdate에서 쓰이는 ui들을
여기서 실행시킬 예정이다

일례로 hitbar, selectedUI, prepareBuilding, resource 등이 있다

활성화 하는 방법은 변화가 있을 경우나 필요한 경우 최신화 되게 할 예정!

+++ 0_Scripts/Project/Survival/01_Attack/2_Missiles/Missile.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public abstract class Missile : MonoBehaviour
 {
 
-    public abstract void Init(Selectable _atker, int _atk, short _prefabIdx);
+    public abstract void Init(Selectable _atker, int _atk, int _prefabIdx);
 
     protected abstract void Used();
