+++ 0_Scripts/Project/Survival/11_Missions/TargetUnit.cs
@@ -25,8 +25,8 @@ public class TargetUnit : Mission
     {
         
         // 유닛 생성 해야한다!
-        ushort selectIdx = target.MyStat.SelectIdx;
-        short prefabIdx = PoolManager.instance.ChkIdx(selectIdx);
+        int selectIdx = target.MyStat.SelectIdx;
+        int prefabIdx = PoolManager.instance.ChkIdx(selectIdx);
 
         curNum = 0;
